add recent repos menu option
